docs(thesis): record login collision fix
